Allow editing max spend limit
In the case where the initial spend limit for the `approve` function
was set to the maximum amount, editing this value would result in the
new limit being silently ignored. The transaction would be submitted
with the original max spend limit.

This occurred because function to generate the new custom data would
look for the expected spend limit in the existing data, then bail if
it was not found. It was not found in these cases because the encoding
we were looking for does not match what was present (the reason has
yet to be deterimined).